complete all functionalities of admin and finish the php work

+++ delete-category.php
@@ -5,10 +5,15 @@ require 'database.php';
 if(isset($_GET['id'])){
     $id = $_GET['id'];
 
+    $update_query = "update posts set category_id =  8 where category_id = $id";
+    $update_result = mysqli_query($conn, $update_query);
 
-    $query = "delete from category where id = $id";
-    $result = mysqli_query($conn, $query);
-    $_SESSION['delete-category-success'] = "category deleted successfully";
+    if(!mysqli_errno($conn)){
+        $query = "delete from category where id = $id";
+        $result = mysqli_query($conn, $query);
+        $_SESSION['delete-category-success'] = "category deleted successfully";
+    }
+
 }
 
 header("location:" .'manage-categories.php');
